fix(#1099): remove dead View ▸ Show Web Inspector command
The command only worked through private WebKit API (_inspector,
_setInspectorAttachmentView:) gated behind #if !ANGLESITE_MAS. Since
direct-download distribution was retired, the shipping Anglesite target
always compiles with ANGLESITE_MAS set, so the code path was permanently
unreachable and the menu item silently did nothing. There's no public
WebKit API to show an embedded inspector, and App Store review rules out
shipping the private one, so the command can never work in the only
build users get.
