+++ Sources/AnglesiteApp/Localizable.xcstrings
@@ -2530,9 +2530,6 @@
     "Show Related Pages" : {
 
     },
-    "Show Web Inspector" : {
-
-    },
     "Show chat panel" : {
 
     },
